botlib: AAS_RocketJumpZVelocity is double, not float; 415 -> 416
The sibling settles it without any disassembly: AAS_BFGJumpZVelocity, the
identical one-line wrapper over AAS_WeaponJumpZVelocity three lines below, is
declared `double` and byte-matches on both oracles.  This one had been narrowed
to `float`, so gcc had to round-trip the callee's ST(0) return through memory
(`fstp DWORD [esp+0xc]; fld DWORD [esp+0xc]`) and grew a 4-byte frame to hold
it.  Q3 has both as `float`; the two shipped binaries do not, and they win.

Also refreshes the README and the memory with the closing figures and the two
generalisations this run earned: an inlined leaf's shape costs its whole call
graph (PC_NameHash 8 rows, ValidClientNumber 6), and where a family of
near-identical functions splits into MATCH and non-MATCH, the matching sibling
is the answer -- diff the declarations before reading disassembly.

Verification:
  loop.sh botlib: 415 -> 416, nothing BROKE; AAS_ShowReachability improved
    158 -> 114 insn_diffs as a side effect.
    byte-identical there at 6 instructions, so the DLL could not have told us
    which type was right.
  make clean && make: no new warning; nm -D shows only the three weak symbols.
  Engine: q2dm1 survives 25 s, clean AAS shutdown.

Source-commit: 7b159459a74d09c5742c9c9c68d317482bf6aa6d

+++ botlib/be_aas_move.c
@@ -248,8 +248,11 @@ double __cdecl AAS_WeaponJumpZVelocity(vec3_t origin, float radiusdamage)
 //----- (1000F750) --------------------------------------------------------
 /* AAS_RocketJumpZVelocity (was sub_1000F750) — Z-velocity from self-rocketing
  * at `origin`; a one-line wrapper over AAS_WeaponJumpZVelocity with the
- * launcher's 120-unit radius damage. */
-float __cdecl AAS_RocketJumpZVelocity(vec3_t origin)
+ * launcher's 120-unit radius damage.  `double`, not `float`: AAS_BFGJumpZVelocity
+ * immediately below is the same one-line wrapper and byte-matches with `double`,
+ * where the `float` narrowing here made gcc round-trip the callee's ST(0) return
+ * through memory (`fstp DWORD; fld DWORD`) and grow a 4-byte frame for it. */
+double __cdecl AAS_RocketJumpZVelocity(vec3_t origin)
 {
   return AAS_WeaponJumpZVelocity(origin, 120.0);
 }
